Fix incorrect link description for empty new.org
When inserting a missing link for the first time, the link description was
incorrectly set to "new.org" rather than the new headline. This is because, during
org-capture-before-finalize-hook, the target buffer is still narrowed, so org did
not think there was a first headline. Now, we widen the target buffer before storing
the headline link.